# laravel11/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SyncBridge API",
        "description": "Production-grade API documentation for SyncBridge integration endpoints.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8080",
            "description": "Primary API server"
        }
    ],
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List products",
                "description": "Returns paginated products with optional filtering and search.",
                "operationId": "productsIndex",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 200,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "synced",
                                "pending",
                                "error",
                                "syncing",
                                "quarantined"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Products fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiSuccessBase"
                                        },
                                        {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 245
                                                },
                                                "page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "rows": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Product"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/products/{productId}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product details",
                "description": "Returns product details with variants and images.",
                "operationId": "productsShow",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiSuccessBase"
                                        },
                                        {
                                            "properties": {
                                                "product": {
                                                    "$ref": "#/components/schemas/Product"
                                                },
                                                "variants": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/ProductVariant"
                                                    }
                                                },
                                                "images": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/ProductImage"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiSuccessBase": {
                "required": [
                    "ok",
                    "message"
                ],
                "properties": {
                    "ok": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "ok"
                    }
                },
                "type": "object"
            },
            "ApiError": {
                "required": [
                    "ok",
                    "error",
                    "errors"
                ],
                "properties": {
                    "ok": {
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "type": "string",
                        "example": "Invalid payload"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "schema": "OpenApiSpec",
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "The field is required."
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "Product": {
                "required": [
                    "id",
                    "ps_id",
                    "name",
                    "sync_status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "ps_id": {
                        "type": "integer",
                        "example": 12345
                    },
                    "name": {
                        "type": "string",
                        "example": "Slim Fit Denim Jacket"
                    },
                    "reference": {
                        "type": "string",
                        "example": "REF-123",
                        "nullable": true
                    },
                    "sync_status": {
                        "type": "string",
                        "example": "synced"
                    },
                    "sync_error": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "ay_style_key": {
                        "type": "string",
                        "example": "BY-STYLE-123",
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 89.99
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ProductVariant": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 7
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "ps_combo_id": {
                        "type": "integer",
                        "example": 1001,
                        "nullable": true
                    },
                    "sku": {
                        "type": "string",
                        "example": "SKU-BLACK-M",
                        "nullable": true
                    },
                    "ean13": {
                        "type": "string",
                        "example": "4006381333931",
                        "nullable": true
                    },
                    "reference": {
                        "type": "string",
                        "example": "REF-BLK-M",
                        "nullable": true
                    },
                    "price_modifier": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "weight": {
                        "type": "number",
                        "format": "float",
                        "example": 0.2
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 5
                    },
                    "color_id": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "size_id": {
                        "type": "integer",
                        "example": 4,
                        "nullable": true
                    },
                    "ay_pushed": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "ProductImage": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 3
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "ps_image_id": {
                        "type": "integer",
                        "example": 9001
                    },
                    "source_url": {
                        "type": "string",
                        "example": "https://cdn.example.com/img/source.jpg",
                        "nullable": true
                    },
                    "public_url": {
                        "type": "string",
                        "example": "https://cdn.example.com/img/public.jpg",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "processed",
                        "nullable": true
                    },
                    "error_message": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "position": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "processed_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Paste SYNCBRIDGE_API_TOKEN as Bearer token.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            },
            "apiTokenHeader": {
                "type": "apiKey",
                "description": "Alternative auth header for sync-protected endpoints.",
                "name": "X-Api-Token",
                "in": "header"
            }
        }
    },
...
}